refactor regex to use rust backend
