+++ rust_regex/include/rust_regex.h
@@ -14,10 +14,20 @@ typedef struct RegMatch {
     const char *endp[10];
 } RegMatch;
 
+typedef struct SearchStat {
+    int cur;
+    int cnt;
+    int exact_match;
+    int incomplete;
+    int last_maxcount;
+} SearchStat;
+
 RegProg* vim_regcomp(const char *pattern, int flags);
 void vim_regfree(RegProg *prog);
 int vim_regexec(RegProg *prog, const char *text, RegMatch *matchp);
 char* vim_regsub(RegProg *prog, const char *text, const char *sub);
+void rust_search_update_stat(const char *pat, const char *text, SearchStat *stat);
+int rust_regex_match(const char *pat, const char *text, int magic, long timeout_ms);
 
 #ifdef __cplusplus
 }
